Style class for single cells in styled tables

This class holds all style settings for a single excel cell. The values are filled in by the user.

# S4 method for STCellStyle
initialize(.Object, fontName, fontHeight,
  fontColor, isBold, isItalic, isStrikeout, underline, boldweight,
  isWrapped, horizontal, vertical, rotation, indent, borderPosition,
  borderColor, borderPen, foregroundColor, backgroundColor, fillPattern,
  dataFormat, isLocked, isHidden, latexVerticalMove, excelPreProcess,
  latexPreProcess, latexFontSize)

Arguments

.Object
A STCellStyle object

fontName
The name of the font that should be used

fontHeight
The font height

fontColor
The font color

isBold
Is bold? TRUE/FALSE

isItalic
Is italic? TRUE/FALSE

isStrikeout
Is striked out? TRUE/FALSE

underline
Numeric weight of the underline

boldweight
Numeric font weight

isWrapped
Is wrappend? TRUE/FALSE

horizontal
String with horizontal alignment

vertical
String with vertical alignment

rotation
Numeric holding the rotation of the text

indent
Numeric holding the indnentation

borderPosition
Character vector with possible border positions

borderColor
Color string (hex or color name)

borderPen
String with border pen value

foregroundColor
Color string for foreground color

backgroundColor
Color string for background color

fillPattern
String with fill pattern

dataFormat
String for Excel-Cell-Data-Format

isLocked
Is Locked? TRUE/FALSE

isHidden
Is Hidden? TRUE/FALSE

latexVerticalMove
A latex string that gives the distance for vertically moving the cell content (e.g.: in order to to vertically center it).

excelPreProcess
A function that can be used to pre process cell values for the Excel table generation

latexPreProcess
A function that can be used to pre process cell values for the LaTeX table generation

latexFontSize
The LaTeX command to set the font size