Style class for single cells in StyledTable objects

This class holds all style settings for a single excel cell. The values are filled in by the user.

# S4 method for StyledCell
initialize(.Object, excel_font_name,
  latex_font_name, excel_font_size, font_color, bold, italic, strikeout,
  underline, excel_boldweight, excel_wrapped, horizontal, vertical,
  rotation, indent, border_position, border_color, excel_border_pen,
  fill_color, excel_background_color, excel_fill_pattern,
  excel_data_format, excel_locked, excel_hidden, latex_vertical_move,
  excel_pre_process, latex_pre_process, latex_font_size)

Arguments

.Object
A StyledCell object

excel_font_name
The name of the font that should be used for the Excel table generation

latex_font_name
The name of the font that should be used for the LaTeX table generation

excel_font_size
The font height

font_color
The font color

bold
Is bold? TRUE/FALSE

italic
Is italic? TRUE/FALSE

strikeout
Is striked out? TRUE/FALSE

underline
Numeric weight of the underline

excel_boldweight
Numeric font weight

excel_wrapped
Is wrappend? TRUE/FALSE

horizontal
String with horizontal alignment

vertical
String with vertical alignment

rotation
Numeric holding the rotation of the text

indent
Numeric holding the indnentation

border_position
Character vector with possible border positions

border_color
Color string (hex or color name)

excel_border_pen
String with border pen value

fill_color
Color string for foreground color

excel_background_color
Color string for background color

excel_fill_pattern
String with fill pattern

excel_data_format
String for Excel-Cell-Data-Format

excel_locked
Is Locked? TRUE/FALSE

excel_hidden
Is Hidden? TRUE/FALSE

latex_vertical_move
A latex string that gives the distance for vertically moving the cell content (e.g.: in order to to vertically center it).

excel_pre_process
A function that can be used to pre process cell values for the Excel table generation

latex_pre_process
A function that can be used to pre process cell values for the LaTeX table generation

latex_font_size
The LaTeX command to set the font size