Create a styled cross table body

This function creates the styled cross table body from a given data.frame. This function works best in combination with the method 'createStyledCrossTableHeader'. For the the creation of the styled cross table body the columns specified in 'subTableCols' are used to separate the table body into several sub tables, where each sub table has the current value in the column defined by 'subTableCols' as sub heading. The columns specified in 'yCols', 'xCols' and 'valueCols' will be used for the cross table calculation. If the table should not be a cross table, but a normal table then the arguments 'xCols' and 'valueCols should be omitted.

createStyledCrossTableBody(data, ...)

# S4 method for data.frame
createStyledCrossTableBody(data,
  subTableCols = NULL, subHeadingStylings = NULL, bodyStyling = NULL,
  yCols, yColsRight = NULL, xCols = NULL, valueCols = NULL,
  valueColStylings = NULL, dropMissingRows = TRUE,
  dropMissingCols = FALSE, fun.aggregate = NULL, fillValues = NULL)

Arguments

data
A data.frame that should be used for the creation of the cross table.

...
Various arguments

subTableCols
(optional) A vector of column names that should be used in order to separate the data.frame into several sub tables, that will be concatenated vertically. Each sub table will have a sub heading row above the sub table body. The sub heading text is just the corresponding value in the column defined by 'subTableCols'. If no sub heading should be introduced at all, the argument 'subTableCols' should be omitted. If the 'subTableCols' is vector with more than one column name, then a hirachy level is introduced for the sub headings, where the first column has the highest and the column given by the last entry in 'subTableCols' has the lowest hirachy level.

subHeadingStylings
(optional) A list of styling functions. A styling function is a function that takes a styled table as its only argument and returns a styled function. The first styling function is applied to the level-1 sub heading (the sub heading defined by the first entry in 'subTableCols'), the second styling function is applied to the level-2 sub heading etc. Alternatively a single styling function can be passed into 'subHeadingStylings', in this case this styling function is applied to all sub heading levels.

bodyStyling
(optional) A styling function that is applied to the generated styled body of each sub table. A styling function is a function that takes a styled table as its only argument and returns a styled function. In order to format the value columns of the resulting cross table use the argument 'valueColStylings'.

yCols
A vector of column names that are unchanged by the cross table compuation. If the table should not be a cross table, but a normal table this argument can be omitted.

yColsRight
A vector of column names that are unchanged by the cross table compuation. This columns are of the same kind as the columns defined by the argument yCols, but aligned to the right of the generated cross table.

xCols
(optional) A vector of column names that are used for column name generation by the cross table compuation. If the table should not be a cross table, but a normal table this argument can be omitted.

valueCols
(optional) A vector of column names that are used as value columns by the cross table compuation. Normally this is just one column. If it is a vector with several column names, then the column names of these columns introduce an extra level in the cross table column hirachy.

valueColStylings
(optional) A single column styling function or a list of column styling functions that has the same length as 'valueCols'. Each column styling function is applied to the corresponding value column (defined to the corresponding column defined by the 'valueCols' argument) of the resulting styled cross table. If a single column styling function is passed, then the same column styling function is applied to all value columns defined in 'valueCols'. If no value column specific styling should be applied, then the 'valueColStylings' argument should be omitted. If the resulting table should not be a cross table at all, but a normal table, then this argument should also be omitted. A column styling function is a function of the type: function(object, cols) ..., which returns a styled table, where 'object' is a styled table and 'cols' is a numeric vector of column numbers.

dropMissingRows
(optional) If set to FALSE and the 'subTableCols' or 'yCols' are sparse factors then the missing factor levels introduce new rows filled filled with NAs. Otherwise no extra rows will be generated. Default value is 'TRUE'.

dropMissingCols
(optional) If set to FALSE and the 'xCols' are sparse factors then the missing factor levels introduce extra cross table columns filled with NA's. Otherwise no extra columns will be generated. Default value is 'FALSE'.

fun.aggregate
(optional) Defines the function that will be used for the cross table cell aggregation. If not defined, the default aggregation function of the 'data.table::dcast' function will be used.

fillValues
(optional) A list of values that is used to fill the missing values in the value columns in the calculated cross table. The first list value is used for the first value column, the second value for the second value column. Hence, the list given in fillValues should have the same length as the valueCols and the types of the fill values should be the same as the corresponding value columns in the data data.frame. If the argument is omitted, then only NA is filled in for the missing cross table values. Caution: If an Excel file is generated from this cross table, NA values will be replaced by empty strings.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:data.table’:
#> 
#>     between, first, last

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# prepare data set for cross table
studentsData <- data.frame(
    country = rep(c("Germany", "Austria"), each = 16),
    year = c(rep("2010", 8), rep("2011", 8)),
    subject = rep(c(rep("Mathematics", 4), rep("Statistics", 4)), 2),
    gender = rep(c(rep("Male", 2), rep("Female", 2)), 4),
    result = rep(c("positive", "negative"), 16),
    N = sample(1:1000, 32)
  ) %>%
    group_by(country, year, subject, gender) %>%
    mutate(rel = 100 * N / sum(N)) %>%
    ungroup
# setup styled header
sHeader <- createStyledCrossTableHeader(
  yColHeadings = c("Year", "Subject"),
  crossTableHeading = "Comparison of test results",
  c("Male", "Female"),
  c("Positive", "Negative"),
  c("count", "in %")
) %>%
  formatStatHeader
# setup styled cross table body
sBody <- createStyledCrossTableBody(
  data = studentsData,
  subTableCols = "country",
  subHeadingStyling = function(object) setHorizontal(object, "center"),
  bodyStyling = function(object) setHorizontal(object, "center"),
  yCols = c("year", "subject"),
  xCols = c("gender", "result"),
  valueCols = c("N", "rel"),
  valueColStylings = list(
    formatStatAbsolute,
    formatStatRelative
  ),
  fillValues = list(0L, 0)
)
# Concat styled header and styled body into a single styled table
sTabel <- styledTable(
  sHeader,
  sBody
)