Method createStyledCrossTableHeader

This function creates a styled header for a cross table. It is best used together with 'createStyledCrossTableBody'. It assumes that the cross table consists of several Y columns that come first and has nested cross table part. That means that the cross table X columns can have more than one header level. For example: Male > Positive | Male > Negative | Female > Positive | Female > Negative In this case the 'Male' and 'Female' headings will be in merged cells above the 'Positive' and 'Negative' cells.

createStyledCrossTableHeader(yColHeadings, ...)

# S4 method for character
createStyledCrossTableHeader(yColHeadings,
  crossTableHeading = NULL, ...)

Arguments

yColHeadings
A character vector that holds the header texts for the Y columns of the cross table. If no other arguments are given the resulting is not a cross table, but a normal table.

...
(optional) One or more vectors of character strings that hold the levels of the first/second/third/... cross table X column headings. These levels are used as header texts in the first/second/third/... row of the generated cross table columns. If no character vector is supplied, then the resulting table is not a cross table, but a normal table.

crossTableHeading
(optional) A single character string that is printed in a merged cell above the Y cross table cols.

Examples

library(dplyr)
# prepare data set for cross table
studentsData <- data.frame(
    country = rep(c("Germany", "Austria"), each = 16),
    year = c(rep("2010", 8), rep("2011", 8)),
    subject = rep(c(rep("Mathematics", 4), rep("Statistics", 4)), 2),
    gender = rep(c(rep("Male", 2), rep("Female", 2)), 4),
    result = rep(c("positive", "negative"), 16),
    N = sample(1:1000, 32)
  ) %>%
    group_by(country, year, subject, gender) %>%
    mutate(rel = 100 * N / sum(N)) %>%
    ungroup
# setup styled header
sHeader <- createStyledCrossTableHeader(
  yColHeadings = c("Year", "Subject"),
  crossTableHeading = "Comparison of test results",
  c("Male", "Female"),
  c("Positive", "Negative"),
  c("count", "in %")
) %>%
  formatStatHeader
# setup styled cross table body
sBody <- createStyledCrossTableBody(
  data = studentsData,
  subTableCols = "country",
  subHeadingStyling = function(object) setHorizontal(object, "center"),
  bodyStyling = function(object) setHorizontal(object, "center"),
  yCols = c("year", "subject"),
  xCols = c("gender", "result"),
  valueCols = c("N", "rel"),
  valueColStylings = list(
    formatStatAbsolute,
    formatStatRelative
  ),
  fillValues = list(0L, 0)
)
# Concat styled header and styled body into a single styled table
sTabel <- styledTable(
  sHeader,
  sBody
)